outcome generates and saves to state all by RNG
unfortunately, I had to use a loop to get the outcome by id from props.outcomes. This is fine but to increase performance, I will consider moving to using a dispatch for getting the outcome instead of getting all outcomes when the component mounts and working with the outcomes in a prop.